# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/catalog": {
      "get": {
        "operationId": "DomainController_catalogAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Catalogue des modules opérationnels (inventaire, triage, ORSEC, …)",
        "tags": [
          "domain"
        ]
      }
    },
    "/api/incident-types": {
      "get": {
        "operationId": "DomainController_incidentTypesList",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Catalogue paramétrable des types d'incident (libellés FR/AR/EN + icônes)",
        "tags": [
          "domain"
        ]
      },
      "post": {
        "operationId": "DomainController_registerIncidentType",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterIncidentTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enregistrer un nouveau type d'incident (Super Admin, audité)",
        "tags": [
          "domain"
        ]
      }
    },
    "/api/dashboard/stats": {
      "get": {
        "operationId": "DomainController_dashboardStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Statistiques de commandement : évolution 30 j, gravité, bilan humain, saturation hospitalière, posture des unités",
        "tags": [
          "domain"
        ]
      }
    },
    "/api/incidents": {
      "get": {
        "operationId": "DomainController_incidents",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Liste des incidents",
        "tags": [
          "domain"
        ]
      },
      "post": {
        "operationId": "DomainController_createIncident",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIncidentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Déclarer un incident (audité) — type validé contre le catalogue",
        "tags": [
          "domain"
        ]
      }
    },
    "/api/incidents/{id}": {
      "patch": {
        "operationId": "DomainController_updateIncident",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIncidentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Modifier ou archiver un incident (audité)",
        "tags": [
          "domain"
        ]
      }
    },
    "/api/sub-incident-types": {
      "get": {
        "operationId": "DomainController_subIncidentTypesList",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Catalogue des sous-types + mapping par type d'incident principal",
        "tags": [
          "domain"
        ]
      }
    },
    "/api/incidents/{id}/sub-incidents": {
      "post": {
        "operationId": "DomainController_addSubIncident",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubIncidentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rattacher un sous-incident (aléa secondaire) à un incident (audité)",
        "tags": [
          "domain"
        ]
      }
    },
    "/api/incidents/{id}/sub-incidents/{subId}": {
      "delete": {
        "operationId": "DomainController_removeSubIncident",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Détacher un sous-incident (audité)",
        "tags": [
          "domain"
        ]
      }
    },
    "/api/units": {
      "get": {
        "operationId": "DomainController_units",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Liste des unités",
        "tags": [
          "domain"
        ]
      },
      "post": {
        "operationId": "DomainController_createUnit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUnitDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créer une unité (audité)",
        "tags": [
          "domain"
        ]
      }
    },
    "/api/units/{id}": {
      "patch": {
        "operationId": "DomainController_updateUnit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUnitDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mettre à jour une unité — un responsable ne peut agir que sur la sienne",
        "tags": [
          "domain"
        ]
      }
    },
    "/api/shelters": {
      "get": {
        "operationId": "DomainController_shelters",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Liste des abris d'hébergement",
        "tags": [
          "domain"
        ]
      }
    },
    "/api/shelters/{id}": {
      "patch": {
        "operationId": "DomainController_updateShelter",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateShelterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mettre à jour un abri — un responsable ne peut agir que sur le sien",
        "tags": [
          "domain"
        ]
      }
    },
    "/api/equipment-parks/{id}/items": {
      "get": {
        "operationId": "DomainController_parkItems",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Parc d'équipement d'une unité",
        "tags": [
          "domain"
        ]
      },
      "post": {
        "operationId": "DomainController_addParkItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEquipDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ajouter un article — dans SON parc uniquement",
        "tags": [
          "domain"
        ]
      }
    },
    "/api/equipment-parks/{id}/items/{eid}": {
      "patch": {
        "operationId": "DomainController_updateParkItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEquipDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Modifier un article — dans SON parc uniquement",
        "tags": [
          "domain"
        ]
      },
      "delete": {
        "operationId": "DomainController_removeParkItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sortir un article du parc — dans SON parc uniquement",
        "tags": [
          "domain"
        ]
      }
    },
    "/api/morgues": {
      "get": {
        "operationId": "DomainController_morgues",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sites mortuaires",
        "tags": [
          "domain"
        ]
      }
    },
    "/api/morgues/{id}": {
      "patch": {
        "operationId": "DomainController_updateMorgue",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMorgueDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mettre à jour un site mortuaire — le sien uniquement",
        "tags": [
          "domain"
        ]
      }
    },
    "/api/morgues/{id}/records": {
      "get": {
        "operationId": "DomainController_mortuaryRecords",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Registre d'identification d'un site mortuaire",
        "tags": [
          "domain"
        ]
      },
      "post": {
        "operationId": "DomainController_admitBody",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdmitBodyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Admettre un corps sous référence provisoire — dans SON site uniquement",
        "tags": [
          "domain"
        ]
      }
    },
...
    "/api/field-hospitals": {
      "get": {
        "operationId": "DomainController_fieldHospitals",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Hôpitaux de campagne déployés",
        "tags": [
          "domain"
        ]
      }
    },
    "/api/feed": {
      "get": {
        "operationId": "DomainController_feed",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fil des événements",
        "tags": [
          "domain"
        ]
      }
    },
...
  },
  "info": {
    "title": "ARGOS API",
    "description": "API du poste de commandement ARGOS — Phase 0/1 : IAM/RBAC (default-deny), audit chaîné, feature flags.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegisterIncidentTypeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "sandstorm",
            "description": "Identifiant (slug)"
          },
          "labels": {
            "$ref": "#/components/schemas/IncidentTypeLabelsDto"
          },
          "icon": {
            "type": "string",
            "description": "Tracé SVG 24×24 (icône en trait)"
          }
        },
        "required": [
          "id",
          "labels"
        ]
      },
...
      "UpdateUnitDto": {
        "type": "object",
        "properties": {
          "cmdt": {
            "type": "string",
            "description": "Commandant de l'unité"
          },
          "eff": {
            "type": "number",
            "minimum": 0,
            "description": "Effectif"
          },
          "dispo": {
            "type": "string",
            "enum": [
              "ready",
              "deployed",
              "standby"
            ],
            "description": "Posture"
          },
          "readiness": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "description": "Taux de préparation (%)"
          }
        }
      },
...
    }
  }
}